requirements.txt
flask
flask-wtf
wtforms
pandas
openpyxl
rdkit-pypi
openbabel-wheel
gunicorn


